test(Sections): add cases of adding end section
- Sections의 새로운 책임인 종점 구간 추가에 대한 테스트 케이스 추가
