--- src/main/java/nextstep/subway/line/domain/sections/AddSectionPolicy.java
@@ -1,5 +1,0 @@
-package nextstep.subway.line.domain.sections;
-
-public interface AddSectionPolicy {
-    boolean addSection(final Section newSection, final Sections sections);
-}
